--- app/src/main/java/dev/arkbuilders/rate/data/Alias.kt
@@ -1,3 +1,0 @@
-package dev.arkbuilders.rate.data
-
-typealias CurrencyCode = String
