--- app/src/main/java/dev/arkbuilders/rate/data/CurrencyName.kt
@@ -1,6 +1,0 @@
-package dev.arkbuilders.rate.data
-
-data class CurrencyName(
-    val code: CurrencyCode,
-    val name: String,
-)
