+++ app/src/main/java/dev/arkbuilders/rate/data/crypto/CryptoCurrencyRepo.kt
@@ -1,10 +1,10 @@
 package dev.arkbuilders.rate.data.crypto
 
-import dev.arkbuilders.rate.data.CurrencyCode
-import dev.arkbuilders.rate.data.CurrencyName
-import dev.arkbuilders.rate.data.CurrencyRate
-import dev.arkbuilders.rate.data.CurrencyRepo
-import dev.arkbuilders.rate.data.CurrencyType
+import dev.arkbuilders.rate.data.model.CurrencyCode
+import dev.arkbuilders.rate.data.model.CurrencyName
+import dev.arkbuilders.rate.data.model.CurrencyRate
+import dev.arkbuilders.rate.data.model.CurrencyRepo
+import dev.arkbuilders.rate.data.model.CurrencyType
 import dev.arkbuilders.rate.data.network.NetworkStatus
 import dev.arkbuilders.rate.data.db.CurrencyRateLocalDataSource
 import dev.arkbuilders.rate.data.db.FetchTimestampDataSource
