+++ app/src/main/java/dev/arkbuilders/rate/data/model/Alias.kt
@@ -1,0 +1,3 @@
+package dev.arkbuilders.rate.data.model
+
+typealias CurrencyCode = String
