+++ app/src/main/java/dev/arkbuilders/rate/data/model/CurrencyAmount.kt
@@ -1,6 +1,7 @@
-package dev.arkbuilders.rate.data
+package dev.arkbuilders.rate.data.model
 
 import android.os.Parcelable
+import dev.arkbuilders.rate.data.model.CurrencyCode
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
