+++ app/src/main/java/dev/arkbuilders/rate/data/model/CurrencyName.kt
@@ -1,0 +1,8 @@
+package dev.arkbuilders.rate.data.model
+
+import dev.arkbuilders.rate.data.model.CurrencyCode
+
+data class CurrencyName(
+    val code: CurrencyCode,
+    val name: String,
+)
