+++ app/src/main/java/dev/arkbuilders/rate/data/model/CurrencyRepo.kt
@@ -1,6 +1,7 @@
-package dev.arkbuilders.rate.data
+package dev.arkbuilders.rate.data.model
 
 import android.util.Log
+import dev.arkbuilders.rate.data.model.CurrencyCode
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
