+++ app/src/main/java/dev/arkbuilders/rate/data/model/QuickCurrency.kt
@@ -1,4 +1,6 @@
-package dev.arkbuilders.rate.data
+package dev.arkbuilders.rate.data.model
+
+import dev.arkbuilders.rate.data.model.CurrencyCode
 
 class QuickCurrency(
     val code: CurrencyCode,
