+++ app/src/main/java/dev/arkbuilders/rate/data/worker/CurrencyMonitorWorker.kt
@@ -3,11 +3,11 @@ package dev.arkbuilders.rate.data.worker
 import android.content.Context
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
-import dev.arkbuilders.rate.data.CurrencyRate
+import dev.arkbuilders.rate.data.model.CurrencyRate
 import dev.arkbuilders.rate.data.GeneralCurrencyRepo
 import dev.arkbuilders.rate.data.db.PairAlertConditionRepo
 import dev.arkbuilders.rate.di.DIManager
-import dev.arkbuilders.rate.data.PairAlertCondition
+import dev.arkbuilders.rate.data.model.PairAlertCondition
 import dev.arkbuilders.rate.presentation.utils.NotificationUtils
 import javax.inject.Inject
