+++ app/src/main/java/dev/arkbuilders/rate/presentation/addcurrency/AddCurrencyViewModel.kt
@@ -7,11 +7,11 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
-import dev.arkbuilders.rate.data.CurrencyAmount
-import dev.arkbuilders.rate.data.CurrencyName
+import dev.arkbuilders.rate.data.model.CurrencyAmount
+import dev.arkbuilders.rate.data.model.CurrencyName
 import dev.arkbuilders.rate.data.GeneralCurrencyRepo
 import dev.arkbuilders.rate.data.assets.AssetsRepo
-import dev.arkbuilders.rate.data.CurrencyCode
+import dev.arkbuilders.rate.data.model.CurrencyCode
 import kotlinx.coroutines.flow.MutableSharedFlow
 import javax.inject.Inject
 import javax.inject.Singleton
